Initial commit for OpenGL demo
Currently allows WASD/arrow key movement of an object on screen. Have a
bunch of little rectangles drawn in the background with a parallax
movement effect.
There are a bunch of methods for drawing basic shapes.
Works with webgl using gopherjs, but is extremely laggy. Even the
desktop version lags a bit. I need to make the little rectangles get
drawn more efficiently.