Major Refactor
Mostly to allow for multiple shaders. Now there's a singleton for each
type of shader which can be accessed like shader.Basic or
shader.Parallax.
Also added support for taking screenshots and cleaned it all up a bit.
It still needs a lot of work, but at least with the parallax shader I
can render thousands of rectangles at once without killing the framerate
in the browser.